Moved Cart to AuthContext. Makes it easier
