+++ context/ServicesProvider.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useReducer, useRef, useState } from "react";
+import React, { useEffect, useReducer, useState } from "react";
 import { getReviewsData } from "../utils/getReviewsData";
 import { getSellersData } from "../utils/getSellersData";
 import { getServicesData } from "../utils/getServicesData";
@@ -15,10 +15,7 @@ const ServicesProvider = ({ children }) => {
   };
 
   const [state, dispatch] = useReducer(servicesReducer, initialState);
-  const [cart, setCart] = useState([]);
   const [serviceId, setServiceId] = useState();
-
-  const initialRender = useRef(true);
 
   const setServices = (newServices) => {
     dispatch({ type: GET_SERVICES, payload: newServices });
@@ -44,29 +41,12 @@ const ServicesProvider = ({ children }) => {
     fetchData();
   }, [serviceId, dispatch]);
 
-  useEffect(() => {
-    if (initialRender.current) {
-      initialRender.current = false;
-      return;
-    }
-    localStorage.setItem("cart", JSON.stringify(cart));
-  }, [cart]);
-
-  useEffect(() => {
-    if (JSON.parse(localStorage.getItem("cart"))) {
-      const storedCart = JSON.parse(localStorage.getItem("cart"));
-      setCart([...cart, ...storedCart]);
-    }
-  }, []);
-
   const values = {
     services: state.services,
     sellers: state.sellers,
     reviews: state.reviews,
     loading: state.loading,
-    cart,
     dispatch,
-    setCart,
     setServices,
     setReviews,
     setServiceId,
